Move mssql migration util to common module
